requirements.txt
boto3
botocore
confluent-kafka
pyyaml